pyproject.toml
[tool.bandit]
skips = ["B101", "B311", "B308"]
exclude_dirs = ['.venv', 'tests']

[tool.mypy]
python_version = "3.13"
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_return_any = true
ignore_missing_imports = true

[project]
name = "online-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-filter (>=25.2,<26.0)",
    "django-silk (>=5.4.3,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "stripe (>=13.0.1,<14.0.0)",
    "crispy-bootstrap4 (>=2025.6,<2026.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "flower (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "djlint (>=1.36.4,<2.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "bandit (>=1.8.6,<2.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "safety (>=3.6.2,<4.0.0)",
    "coverage (>=7.11.0,<8.0.0)"
]



[tool.poetry]
package-mode = false
